// Portal-RH_Frontend-Servidor/src/app/models/solicitacoes/reducao-carga-horaria/reducao-carga-horaria.model.ts
export interface ReducaoCH {
  matricula: number;
  cpf: string;
  rg: string;
  nome: string;
  localTrabalho: string;
  cargo: string;
  cep: string;
  endereco: string;
  bairro: string;
  municipio: string;
  numeroEndereco: string;
  complementoEndereco: string;
  tipo: string;
  conteudo: string;
}